modify following files pom package name git ignore readme
