+++ .gitignore
@@ -4,3 +4,25 @@
 !**/*.md
 !.gitignore
 !pom.xml
+
+/target
+*.class
+
+# Mobile Tools for Java (J2ME)
+.mtj.tmp/
+
+# Package Files #
+*.jar
+*.war
+*.ear
+
+# virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
+hs_err_pid*
+
+# Eclipse
+/.project
+/.classpath
+
+# Idea
+/redis-replicator.iml
+/.idea
