+++ src/main/java/com/redislabs/modules/rejson/JReJSON.java
@@ -1,4 +1,32 @@
-package io.rejson;
+/*
+ * BSD 2-Clause License
+ *
+ * Copyright (c) 2017, Redis Labs
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *
+ * * Redistributions of source code must retain the above copyright notice, this
+ *   list of conditions and the following disclaimer.
+ *
+ * * Redistributions in binary form must reproduce the above copyright notice,
+ *   this list of conditions and the following disclaimer in the documentation
+ *   and/or other materials provided with the distribution.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+ * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
+ * FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+ * DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+ * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+ * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+ * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+package com.redislabs.modules.rejson;
 
 import com.google.gson.Gson;
 import redis.clients.jedis.Jedis;
@@ -95,7 +123,7 @@ public class JReJSON {
      */
     public static Long del(Jedis conn, String key, Path... path) {
 
-        ArrayList<byte[]> args = new ArrayList(2);
+        List<byte[]> args = new ArrayList(2);
 
         args.add(SafeEncoder.encode(key));
         args.add(SafeEncoder.encode(getSingleOptionalPath(path).toString()));
@@ -117,7 +145,7 @@ public class JReJSON {
      */
     public static Object get(Jedis conn, String key, Path... paths) {
 
-        ArrayList<byte[]> args = new ArrayList(2);
+        List<byte[]> args = new ArrayList(2);
 
         args.add(SafeEncoder.encode(key));
         for (Path p :paths) {
@@ -143,7 +171,7 @@ public class JReJSON {
      */
     public static void set(Jedis conn, String key, Object object, ExistenceModifier flag, Path... path) {
 
-        ArrayList<byte[]> args = new ArrayList(4);
+        List<byte[]> args = new ArrayList(4);
 
         args.add(SafeEncoder.encode(key));
         args.add(SafeEncoder.encode(getSingleOptionalPath(path).toString()));
@@ -178,11 +206,9 @@ public class JReJSON {
      * @param path optional single path in the object, defaults to root
      * @return the Java class of the requested object
      */
-    public static Class<? extends Object> type(Jedis conn, String key, Path... path) {
+    public static Class<?> type(Jedis conn, String key, Path... path) {
 
-        ArrayList<byte[]> args = new ArrayList(2);
-
-
+        List<byte[]> args = new ArrayList(2);
 
         args.add(SafeEncoder.encode(key));
         args.add(SafeEncoder.encode(getSingleOptionalPath(path).toString()));
