+++ src/test/java/com/redislabs/modules/rejson/ClientTest.java
@@ -1,36 +1,60 @@
-package io.rejson;
+/*
+ * BSD 2-Clause License
+ *
+ * Copyright (c) 2017, Redis Labs
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *
+ * * Redistributions of source code must retain the above copyright notice, this
+ *   list of conditions and the following disclaimer.
+ *
+ * * Redistributions in binary form must reproduce the above copyright notice,
+ *   this list of conditions and the following disclaimer in the documentation
+ *   and/or other materials provided with the distribution.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+ * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
+ * FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+ * DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+ * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+ * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+ * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+package com.redislabs.modules.rejson;
 
 import com.google.gson.Gson;
-
 import org.junit.Before;
 import org.junit.Test;
 import redis.clients.jedis.Jedis;
-import redis.clients.jedis.JedisPool;
-import redis.clients.jedis.JedisPoolConfig;
-import redis.clients.jedis.Pipeline;
 
 import static junit.framework.TestCase.*;
 
-class FooBarObject extends Object {
-    public String foo;
-
-    public FooBarObject() {
-        this.foo = "bar";
-    }
-}
-
-/* A simple class that represents an object in real life */
-class IRLObject extends Object {
-    public String str;
-    public boolean bTrue;
-
-    public IRLObject() {
-        this.str = "string";
-        this.bTrue = true;
-    }
-}
-
 public class ClientTest {
+
+    /* A simple class that represents an object in real life */
+    private static class IRLObject {
+        public String str;
+        public boolean bTrue;
+
+        public IRLObject() {
+            this.str = "string";
+            this.bTrue = true;
+        }
+    }
+
+    private static class FooBarObject {
+        public String foo;
+
+        public FooBarObject() {
+            this.foo = "bar";
+        }
+    }
 
     private Gson g;
     private String host="localhost";
